utils: relocate request_auth to utils
When request_auth was only used by admin server it was ok to keep it in
redpanda directory/library. But once it was used in panda proxy this
created a circular dependency causing problems for the meson build.
Moved to utils as a place that resolves the dependency.
