+++ src/v/pandaproxy/rest/proxy.h
@@ -16,8 +16,8 @@
 #include "pandaproxy/rest/configuration.h"
 #include "pandaproxy/server.h"
 #include "pandaproxy/util.h"
-#include "redpanda/request_auth.h"
 #include "seastarx.h"
+#include "utils/request_auth.h"
 
 #include <seastar/core/future.hh>
 #include <seastar/core/sharded.hh>
