+++ src/v/redpanda/admin_server.h
@@ -15,9 +15,9 @@
 #include "coproc/partition_manager.h"
 #include "model/metadata.h"
 #include "pandaproxy/schema_registry/fwd.h"
-#include "request_auth.h"
 #include "rpc/connection_cache.h"
 #include "seastarx.h"
+#include "utils/request_auth.h"
 
 #include <seastar/core/scheduling.hh>
 #include <seastar/core/sstring.hh>
