+++ src/v/utils/request_auth.h
